refactor: utils에 있던 엔티티 관련 함수들을 models로 이동
