+++ src/refactoring/hooks/useCart.ts
@@ -1,7 +1,12 @@
-// useCart.ts
 import { useState } from 'react';
 import { CartItem, Coupon, Product } from '../../types';
-import { calculateCartTotal, updateCartItemQuantity } from '../models/cart';
+import {
+  calculateCartTotal,
+  addCartProduct,
+  updateCartItemQuantity,
+  removeCartProduct,
+  getDefaultCartTotal,
+} from '../models/cart';
 
 // 장바구니 관련 커스텀 훅
 export const useCart = () => {
@@ -44,23 +49,3 @@ export const useCart = () => {
     selectedCoupon,
   };
 };
-
-// util 함수들
-// 상품 추가 또는 수량 업데이트
-const addCartProduct = (cart: CartItem[], product: Product): CartItem[] => {
-  const existingItem = cart.find((item) => item.product.id === product.id);
-  return existingItem
-    ? updateCartItemQuantity(cart, product.id, existingItem.quantity + 1)
-    : [...cart, { product, quantity: 1 }];
-};
-
-// 상품 제거
-const removeCartProduct = (cart: CartItem[], productId: string): CartItem[] =>
-  cart.filter((item) => item.product.id !== productId);
-
-// 빈 장바구니 총액을 반환
-const getDefaultCartTotal = () => ({
-  totalBeforeDiscount: 0,
-  totalAfterDiscount: 0,
-  totalDiscount: 0,
-});
