+++ src/refactoring/models/cart.ts
@@ -1,4 +1,5 @@
-import { CartItem, Coupon } from '../../types';
+import { CartItem, Coupon, Product } from '../../types';
+import { getMaxDiscount } from './discount';
 
 // 단일 상품 가격 계산 (최대 할인 적용)
 export const calculateItemTotal = (item: CartItem): number =>
@@ -11,11 +12,13 @@ export const calculateItemTotal = (item: CartItem): number =>
 const applyDiscount = (amount: number, discountRate: number): number =>
   discountRate > 0 ? amount * (1 - discountRate) : amount;
 
-// 최대 할인율 계산
+// 장바구니 상품의 최대 할인율을 계산하기
 export const getMaxApplicableDiscount = (item: CartItem) =>
-  item.product.discounts
-    .filter((discount) => item.quantity >= discount.quantity)
-    .reduce((max, discount) => Math.max(max, discount.rate), 0);
+  getMaxDiscount(
+    item.product.discounts.filter(
+      (discount) => item.quantity >= discount.quantity,
+    ),
+  );
 
 // 쿠폰 적용 후 장바구니 총액 계산
 // 할인 전 총액, 할인 후 총액, 총 할인액을 계산한다.
@@ -62,6 +65,17 @@ const calculateCouponDiscount = (
     : (totalAfterDiscount * coupon.discountValue) / 100;
 };
 
+// 상품 추가 또는 수량 업데이트
+export const addCartProduct = (
+  cart: CartItem[],
+  product: Product,
+): CartItem[] => {
+  const existingItem = cart.find((item) => item.product.id === product.id);
+  return existingItem
+    ? updateCartItemQuantity(cart, product.id, existingItem.quantity + 1)
+    : [...cart, { product, quantity: 1 }];
+};
+
 // 장바구니 상품들의 수량을 업데이트 후 재고에 반영
 export const updateCartItemQuantity = (
   cart: CartItem[],
@@ -86,3 +100,16 @@ const updateSingleItemQuantity = (
 
   return updatedQuantity > 0 ? { ...item, quantity: updatedQuantity } : null;
 };
+
+// 상품 제거
+export const removeCartProduct = (
+  cart: CartItem[],
+  productId: string,
+): CartItem[] => cart.filter((item) => item.product.id !== productId);
+
+// 빈 장바구니 총액을 반환
+export const getDefaultCartTotal = () => ({
+  totalBeforeDiscount: 0,
+  totalAfterDiscount: 0,
+  totalDiscount: 0,
+});
