+++ src/refactoring/models/discount.ts
@@ -1,0 +1,4 @@
+// 최대 할인율 계산하기
+export const getMaxDiscount = (
+  discounts: { quantity: number; rate: number }[],
+) => discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
